requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
requests==2.32.3
pydantic==2.10.4
PyYAML==6.0.2
